# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend created by Hanif using Gin",
        "title": "Backend Ewallet",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/transaction/receivers": {
            "get": {
                "description": "List other users for transfer purpose with search and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Search and Retriever Receiver List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search keyword by name and phone",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max data per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_ReceiverListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_iamhanif11_ewallet-backend_internal_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.ReceiverListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ReceiverResponse"
                    }
                },
                "pages": {
                    "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.PaginationResponse"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.ReceiverResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "reciver": {
                    "type": "string"
                }
            }
        },
...
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_ReceiverListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ReceiverListResponse"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token used for authorization",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}